Move test to `./cli`
