+++ .github/workflows/ci-base-tests-linux.yml
@@ -23,6 +23,7 @@ jobs:
     strategy:
       matrix:
         tests:
+          - ./cli
           - ./envision
           - ./smarts/contrib
           - ./smarts/core --nb-exec-timeout 65536
