+++ .github/workflows/ci-base-tests-mac.yml
@@ -16,6 +16,7 @@ jobs:
     strategy:
       matrix:
         tests:
+          - ./cli
           - ./envision
           - ./smarts/contrib
           - ./examples/tests --ignore=./examples/tests/test_learning.py
